daily_streaming/requirements.txt
# Daily Streaming Requirements
# Install with: pip install -r requirements.txt

aiohttp>=3.9.0
websockets>=12.0
pipecat-ai>=0.0.50
pillow>=10.0.0

